requirements.txt
amqp
psutil
tablib
pytz
pyyaml
sqlalchemy
mysqlclient
telethon
celery==4.2.0
Django==2.0.7
gunicorn
django-mysql
djangorestframework==3.8.2
requests==2.19.1


